Virtuemart frontview will be deprecated next time. Added plenty of values to the category view xml, still WIP Added currency to payment/shipment method, plugins can easily use it by adding $this->setConvertable(array('valuestoconvert1','toconvert2',...)); added function vmVariables to vmjsapi.php for easier adding of the vm vars. addStandardDefaultViewCommands was ignoring params, fixed added meta data for manufacturers fixed problem with php7 f or setProperties in orders.php added "new" for rating in BE
git-svn-id: https://dev.virtuemart.net/svn/virtuemart/branches/com_virtuemart.3.0.12.4@9200 87cc427b-4a1e-0410-b729-b2260ec0f063

+++ administrator/components/com_virtuemart/helpers/vmviewadmin.php
@@ -100,10 +100,10 @@ class VmViewAdmin extends JViewLegacy {
 		if(vmAccess::manager($view.'.edit')){
 			JToolBarHelper::editList();
 		}
-		if(vmAccess::manager($view.'.create')){
+		if($showNew and vmAccess::manager($view.'.create')){
 			JToolBarHelper::addNew();
 		}
-		if(vmAccess::manager($view.'.delete')){
+		if($showDelete and vmAccess::manager($view.'.delete')){
 			JToolBarHelper::spacer('10');
 			JToolBarHelper::deleteList();
 		}
